cui-call-display: force css on answer and hangup buttons using gtk_style_context_add_provider
GTK3 does not automatically apply CSS to widgets with template IDs
the same way GTK4 does. As a result, the #hang_up and #answer rules in
style.css were ignored and both buttons used default Adwaita theme colors.

TODO: figure out why this is needed and get rid of it

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>